[#172767793] Fix import error from Javier, add int to PEP, add README, Makefile
