+++ Makefile
@@ -1,0 +1,10 @@
+.PHONY: all install test
+
+all: install test
+
+install:
+	pip install -r requirements.txt
+
+test:
+	pytest
+
