+++ README.md
@@ -1,1 +1,32 @@
 # z-lstm
+
+This repository contain the source code and workflow templates to run [Prediction with LSTM](http://orquestra.io/docs/tutorial/lstm/) in Orquestra.
+
+Example workflows can be found inside `example/`.
+
+
+---
+
+## Developing
+
+Create an virtual environment:
+
+```
+virtualenv venv
+source venv/bin/activate
+```
+
+Install depedencies:
+
+```
+make install
+```
+
+### Running tests
+
+Tests are located in `src/lstm/test`. Run them with:
+
+```
+make test
+```
+
