+++ src/python/lstm/test/test_data_artifact.json
