+++ src/python/lstm/test/test_trained_model_specs.json
