+++ src/python/lstm/test/test_untrained_model_specs.json
