--- src/python/zml/lstm/__init__.py
@@ -1,2 +1,0 @@
-from .data import *
-from .model import *
